refactor: move model classes to model dir
