+++ core/src/main/java/org/justserve/model/ProjectEventLocation.java
@@ -1,4 +1,4 @@
-package org.justserve.model.graph;
+package org.justserve.model;
 
 import io.micronaut.core.annotation.Introspected;
 import io.micronaut.serde.annotation.Serdeable;
@@ -9,5 +9,5 @@ import lombok.experimental.Accessors;
 @Accessors(chain = true)
 @Serdeable
 @Introspected
-public class CivicGeography {
+public class ProjectEventLocation {
 }
