+++ core/src/main/java/org/justserve/model/User.java
@@ -1,4 +1,4 @@
-package org.justserve.model.graph;
+package org.justserve.model;
 
 import io.micronaut.serde.annotation.Serdeable;
