+++ core/src/main/java/org/justserve/model/graph/CreateEventVariables.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
+import org.justserve.model.ProjectEvent;
 
 import java.util.UUID;
